Refactor clock view model.
